+++ sfsimodels/models/__init__.py
@@ -2,8 +2,9 @@ from sfsimodels.models.hazards import SeismicHazard
 from sfsimodels.models.foundations import Foundation, FoundationPad, FoundationRaft
 from sfsimodels.models.soils import Soil, CriticalSoil, StressDependentSoil, SoilCritical, SoilProfile
 from sfsimodels.models.buildings import Building, \
-    Section, SDOFBuilding, NullBuilding
-from sfsimodels.models import material
+    SDOFBuilding, NullBuilding
+from sfsimodels.models.sections import Section
+from sfsimodels.models import materials
 from sfsimodels.models.systems import SoilStructureSystem
 
 from sfsimodels.models.buildings import FrameBuilding, WallBuilding, FrameBuilding2D  # deprecated objects